@@ -23,19 +23,20 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.web.SecurityFilterChain;
 
 @Profile("local")
 @Configuration
-public class LocalOauthSecurityConfig extends WebSecurityConfigurerAdapter {
+public class LocalOauthSecurityConfig {
 
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http
+    @Bean
+    protected SecurityFilterChain configure(HttpSecurity http) throws Exception {
+        return http
                 .authorizeRequests(auth -> auth
                         .anyRequest().permitAll())
                 .csrf().disable()
-                .headers().frameOptions().disable();
+                .headers().frameOptions().disable().and().build();
     }
 
     @Bean
